Update for Kangaroo -Added more comments -Added function to requisite .h file -Also made some edits to syntax and to the comments for the encoder buffer code files.
